Rearrange code for more overview
