+++ lib/util/forecast_converter.py
@@ -1,58 +1,16 @@
-import math
 from datetime import datetime
-from typing import Union, Tuple
+from typing import Union
 
-from lib.homeassistant import (
-    HomeAssistantAdapter, RequestError, HomeAssistantForecast, HomeAssistantHourlyForecast, HomeAssistantDailyForecast,
-    HomeAssistantWeatherCondition, HomeAssistantForecastMeta, HomeAssistantSunInfo
-)
-from lib.kodi import (
-    KodiLogLevel, KodiGeneralForecastData, KodiForecastData,
-    KodiCurrentForecastData, KodiWindDirectionCode, KodiConditionCode, KodiHourlyForecastData, KodiDailyForecastData
-)
-from lib.unit.speed import SpeedKph, SpeedUnits, Speed
-from lib.unit.temperature import TemperatureUnits, TemperatureCelsius, Temperature
-from lib.weather_plugin_adapter import KodiHomeAssistantWeatherPluginAdapter, HomeAssistantWeatherPluginStrings
+from lib.homeassistant import HomeAssistantHourlyForecast, HomeAssistantForecastMeta, HomeAssistantDailyForecast, \
+    HomeAssistantForecast, HomeAssistantSunInfo, HomeAssistantWeatherCondition
+from lib.kodi import KodiHourlyForecastData, KodiWindDirectionCode, KodiDailyForecastData, KodiForecastData, \
+    KodiGeneralForecastData, KodiCurrentForecastData, KodiConditionCode
+from lib.unit.speed import SpeedUnits
+from lib.unit.temperature import TemperatureUnits
+from lib.util.thermal_comfort import ThermalComfort
 
 
-class KodiHomeAssistantWeatherPlugin:
-    def __init__(self):
-        self._kodi_adapter = KodiHomeAssistantWeatherPluginAdapter()
-        self._kodi_adapter.log("Home Assistant Weather started.")
-
-        if not self._kodi_adapter.required_settings_done():
-            self._kodi_adapter.dialog(message_id=HomeAssistantWeatherPluginStrings.SETTINGS_REQUIRED)
-            self._kodi_adapter.log("Settings for Home Assistant Weather not yet provided. Plugin will not work.")
-        else:
-            self.apply_forecast()
-        self._kodi_adapter.log("Home Assistant Weather init finished.")
-
-    def _get_forecast_handling_errors(self) -> Tuple[HomeAssistantForecast, HomeAssistantSunInfo]:
-        try:
-            return (
-                HomeAssistantAdapter.get_forecast(
-                    server_url=self._kodi_adapter.home_assistant_url,
-                    entity_id=self._kodi_adapter.home_assistant_entity_forecast,
-                    token=self._kodi_adapter.home_assistant_token
-                ),
-                HomeAssistantAdapter.get_sun_info(
-                    server_url=self._kodi_adapter.home_assistant_url,
-                    entity_id=self._kodi_adapter.home_assistant_entity_sun,
-                    token=self._kodi_adapter.home_assistant_token
-                )
-            )
-        except RequestError as e:
-            self._kodi_adapter.log(
-                message=f"Could not retrieve forecast from Home Assistant: {e.error_code}", level=KodiLogLevel.ERROR
-            )
-            if e.error_code == 401:
-                message = HomeAssistantWeatherPluginStrings.HOMEASSISTANT_UNAUTHORIZED
-            elif e.error_code == -1:
-                message = HomeAssistantWeatherPluginStrings.HOMEASSISTANT_UNREACHABLE
-            else:
-                message = HomeAssistantWeatherPluginStrings.HOMEASSISTANT_UNEXPECTED_RESPONSE
-            self._kodi_adapter.dialog(message_id=message)
-
+class ForecastConverter:
     @staticmethod
     def __translate_hourly_ha_forecast_to_kodi_forecast(
             ha_forecast: HomeAssistantHourlyForecast, forecast_meta: HomeAssistantForecastMeta,
@@ -60,24 +18,24 @@ class KodiHomeAssistantWeatherPlugin:
     ) -> KodiHourlyForecastData:
         temperature = TemperatureUnits[forecast_meta.temperature_unit](ha_forecast.temperature)
         wind_speed = SpeedUnits[forecast_meta.wind_speed_unit](ha_forecast.wind_speed)
-        timestamp = KodiHomeAssistantWeatherPlugin.__parse_homeassistant_datetime(datetime_str=ha_forecast.datetime)
+        timestamp = ForecastConverter.__parse_homeassistant_datetime(datetime_str=ha_forecast.datetime)
         return KodiHourlyForecastData(
             temperature=temperature,
             wind_speed=wind_speed,
             wind_direction=KodiWindDirectionCode.from_bearing(bearing=ha_forecast.wind_bearing),
-            precipitation=KodiHomeAssistantWeatherPlugin.__format_precipitation(
+            precipitation=ForecastConverter.__format_precipitation(
                 precipitation=ha_forecast.precipitation, precipitation_unit=forecast_meta.precipitation_unit
             ),
             humidity=ha_forecast.humidity,
-            feels_like=KodiHomeAssistantWeatherPlugin.__calculate_feels_like(
+            feels_like=ThermalComfort.feels_like(
                 temperature=temperature,
                 wind_speed=wind_speed
             ),
-            dew_point=KodiHomeAssistantWeatherPlugin.__calculate_dew_point(
+            dew_point=ThermalComfort.dew_point(
                 temperature=temperature,
                 humidity_percent=ha_forecast.humidity
             ),
-            condition=KodiHomeAssistantWeatherPlugin.__translate_condition(
+            condition=ForecastConverter.__translate_condition(
                 ha_condition=ha_forecast.condition, is_night=not (sunrise.time() < timestamp.time() < sunset.time())
             ),
             timestamp=timestamp,
@@ -94,56 +52,57 @@ class KodiHomeAssistantWeatherPlugin:
             temperature=temperature,
             wind_speed=wind_speed,
             wind_direction=KodiWindDirectionCode.from_bearing(bearing=ha_forecast.wind_bearing),
-            precipitation=KodiHomeAssistantWeatherPlugin.__format_precipitation(
+            precipitation=ForecastConverter.__format_precipitation(
                 precipitation=ha_forecast.precipitation, precipitation_unit=forecast_meta.precipitation_unit
             ),
-            condition=KodiHomeAssistantWeatherPlugin.__translate_condition(ha_condition=ha_forecast.condition),
-            timestamp=KodiHomeAssistantWeatherPlugin.__parse_homeassistant_datetime(datetime_str=ha_forecast.datetime),
+            condition=ForecastConverter.__translate_condition(ha_condition=ha_forecast.condition),
+            timestamp=ForecastConverter.__parse_homeassistant_datetime(datetime_str=ha_forecast.datetime),
             low_temperature=low_temperature,
         )
 
-    def __translate_ha_forecast_to_kodi_forecast(
-            self, ha_forecast: HomeAssistantForecast, ha_sun_info: HomeAssistantSunInfo) -> KodiForecastData:
+    @staticmethod
+    def translate_ha_forecast_to_kodi_forecast(
+            ha_forecast: HomeAssistantForecast, ha_sun_info: HomeAssistantSunInfo) -> KodiForecastData:
         temperature = TemperatureUnits[ha_forecast.current.temperature_unit](ha_forecast.current.temperature)
         wind_speed = SpeedUnits[ha_forecast.current.wind_speed_unit](ha_forecast.current.wind_speed)
-        sunrise = KodiHomeAssistantWeatherPlugin.__parse_homeassistant_datetime(ha_sun_info.next_rising)
-        sunset = KodiHomeAssistantWeatherPlugin.__parse_homeassistant_datetime(ha_sun_info.next_setting)
+        sunrise = ForecastConverter.__parse_homeassistant_datetime(ha_sun_info.next_rising)
+        sunset = ForecastConverter.__parse_homeassistant_datetime(ha_sun_info.next_setting)
         return KodiForecastData(
             General=KodiGeneralForecastData(
-                location=self._kodi_adapter.override_location or ha_forecast.current.friendly_name,
+                location=ha_forecast.current.friendly_name,
                 attribution=ha_forecast.current.attribution,
             ),
             Current=KodiCurrentForecastData(
                 temperature=temperature,
                 wind_speed=wind_speed,
                 wind_direction=KodiWindDirectionCode.from_bearing(bearing=ha_forecast.current.wind_bearing),
-                precipitation=KodiHomeAssistantWeatherPlugin.__format_precipitation(
+                precipitation=ForecastConverter.__format_precipitation(
                     precipitation=ha_forecast.hourly[0].precipitation if len(ha_forecast.hourly) > 0 else None,
                     precipitation_unit=ha_forecast.current.precipitation_unit
-                ),                                              # conversion not implemented in Kodi
-                condition=KodiHomeAssistantWeatherPlugin.__translate_condition(
+                ),  # conversion not implemented in Kodi
+                condition=ForecastConverter.__translate_condition(
                     ha_condition=ha_forecast.hourly[0].condition if len(ha_forecast.hourly) > 0 else None,
                     is_night=not (sunrise.time() < datetime.now().time() < sunset.time())
                 ),
                 humidity=ha_forecast.current.humidity,
-                feels_like=KodiHomeAssistantWeatherPlugin.__calculate_feels_like(
+                feels_like=ThermalComfort.feels_like(
                     temperature=temperature,
                     wind_speed=wind_speed,
                 ),
-                dew_point=KodiHomeAssistantWeatherPlugin.__calculate_dew_point(
+                dew_point=ThermalComfort.dew_point(
                     temperature=temperature,
                     humidity_percent=ha_forecast.current.humidity
                 ),
                 uv_index=int(ha_forecast.current.uv_index),
                 cloudiness=int(ha_forecast.current.cloud_coverage),
-                pressure=KodiHomeAssistantWeatherPlugin.__format_pressure(
+                pressure=ForecastConverter.__format_pressure(
                     pressure=ha_forecast.current.pressure, pressure_unit=ha_forecast.current.pressure_unit
                 ),
                 sunrise=sunrise,
                 sunset=sunset,
             ),
             HourlyForecasts=[
-                self.__translate_hourly_ha_forecast_to_kodi_forecast(
+                ForecastConverter.__translate_hourly_ha_forecast_to_kodi_forecast(
                     ha_forecast=hourly_forecast,
                     forecast_meta=ha_forecast.current,
                     sunrise=sunrise,
@@ -152,39 +111,12 @@ class KodiHomeAssistantWeatherPlugin:
                 for hourly_forecast in ha_forecast.hourly
             ],
             DailyForecasts=[
-                self.__translate_daily_ha_forecast_to_kodi_forecast(
+                ForecastConverter.__translate_daily_ha_forecast_to_kodi_forecast(
                     ha_forecast=daily_forecast,
                     forecast_meta=ha_forecast.current
                 )
                 for daily_forecast in ha_forecast.daily
             ]
-        )
-
-    @staticmethod
-    def __calculate_dew_point(temperature: Temperature, humidity_percent: float) -> TemperatureCelsius:
-        temperature_celsius = TemperatureCelsius.from_si_value(temperature.si_value())
-        # obtain saturation vapor pressure (pressure at which water in air will condensate)
-        vapor_pressure_sat = 6.11 * 10.0 ** (7.5 * temperature_celsius.value / (237.7 + temperature_celsius.value))
-        # we set a minimum of .075 % to make the math defined
-        humidity_percent = max(humidity_percent, 0.075)
-        # calculate actual vapor pressure (water in air will condensate at approx. 100 % humidity), linear correlation
-        vapor_pressure_act = (humidity_percent * vapor_pressure_sat) / 100
-        # Now you are ready to use the following formula to obtain the dewpoint temperature.
-        return TemperatureCelsius(
-            (-430.22 + 237.7 * math.log(vapor_pressure_act)) / (-math.log(vapor_pressure_act) + 19.08)
-        )
-
-    @staticmethod
-    def __calculate_feels_like(temperature: Temperature, wind_speed: Speed) -> TemperatureCelsius:
-        temperature_celsius = TemperatureCelsius.from_si_value(temperature.si_value())
-        wind_speed_kph = SpeedKph.from_si_value(wind_speed.si_value())
-        # Model: Wind Chill JAG/TI Environment Canada
-        # see https://en.wikipedia.org/wiki/Wind_chill#North_American_and_United_Kingdom_wind_chill_index
-        return TemperatureCelsius(
-                + 13.12
-                + 0.6215 * temperature_celsius.value
-                - 11.37 * wind_speed_kph.value ** 0.16
-                + 0.3965 * temperature_celsius.value * wind_speed_kph.value ** 0.16
         )
 
     @staticmethod
@@ -247,20 +179,3 @@ class KodiHomeAssistantWeatherPlugin:
     def __parse_homeassistant_datetime(datetime_str: str) -> datetime:
         # tz=None adds time offset to match Kodi's set time
         return datetime.fromisoformat(datetime_str).astimezone(tz=None)
-
-    def apply_forecast(self):
-        forecast, sun_info = self._get_forecast_handling_errors()
-        if forecast is None:
-            self._kodi_adapter.log(message="No forecasts were found.", level=KodiLogLevel.WARNING)
-            self._kodi_adapter.clear_weather_properties()
-            return
-        if sun_info is None:
-            self._kodi_adapter.log(message="No sun info was found.", level=KodiLogLevel.WARNING)
-            self._kodi_adapter.clear_weather_properties()
-            return
-        kodi_forecast = self.__translate_ha_forecast_to_kodi_forecast(
-            ha_forecast=forecast,
-            ha_sun_info=sun_info,
-        )
-        self._kodi_adapter.set_weather_properties(forecast=kodi_forecast)
-        self._kodi_adapter.log(message="Weather updated successfully.", level=KodiLogLevel.INFO)
